Create 2 FSR SENSOR & IMU 6050 WITH ESP32
This project involves developing a sensorized walking cane to assist elderly
individuals by monitoring their health and mobility. It uses two FSR sensors to
detect the pressure exerted on the cane, which helps in monitoring wound
healing. Additional sensors, including a PPG sensor and MPU6050, are used for
vital sign monitoring and tracking orientation and acceleration. The system is
equipped with an RTC for time tracking and uses an ESP32 for wireless
communication. The aim is to create a device that provides continuous health
monitoring while ensuring ease of use for the elderly.